Move iter_over_entity_members from scenario to builder
