+++ openfisca_core/simulation_builder.py
@@ -12,7 +12,6 @@ from openfisca_core.commons import basestring_type
 from openfisca_core.errors import VariableNotFound, SituationParsingError, PeriodMismatchError
 from openfisca_core.indexed_enums import Enum
 from openfisca_core.periods import key_period_size, period as make_period
-from openfisca_core.scenarios import iter_over_entity_members
 from openfisca_core.simulations import Simulation
 from openfisca_core.tools import eval_expression
 
@@ -342,3 +341,24 @@ def clean_person_list(data):
     if isinstance(data, list):
         return [str(item) if isinstance(item, int) else item for item in data]
     return data
+
+
+def iter_over_entity_members(entity_description, scenario_entity):
+    # One by one, yield individu_role, individy_legacy_role, individu_id
+    legacy_role_i = 0
+    for role in entity_description.roles:
+        role_name = role.plural or role.key
+        individus = scenario_entity.get(role_name)
+
+        if individus:
+            if not type(individus) == list:
+                individus = [individus]
+
+            legacy_role_j = 0
+            for individu in individus:
+                if role.subroles:
+                    yield role.subroles[legacy_role_j], legacy_role_i + legacy_role_j, individu
+                else:
+                    yield role, legacy_role_i + legacy_role_j, individu
+                legacy_role_j += 1
+        legacy_role_i += (role.max or 1)
